Add Collect Target Renderers Pass
