+++ Editor/NDMF/Passes.CollectTargetRenderersPass.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d64e342406c12a548bd5678bbbfdd57b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
